+++ apps/client/src/app/adapter/custom-date-adapter.ts
@@ -1,9 +1,10 @@
 import { getDateFormatString } from '@ghostfolio/common/helper';
 
-import { inject } from '@angular/core';
+import { inject, Service } from '@angular/core';
 import { MAT_DATE_LOCALE, NativeDateAdapter } from '@angular/material/core';
 import { addYears, format, getYear, parse } from 'date-fns';
 
+@Service({ autoProvided: false })
 export class CustomDateAdapter extends NativeDateAdapter {
   public override locale = inject<string>(MAT_DATE_LOCALE);