+++ libs/common/src/lib/types/user-with-settings.type.ts
@@ -13,6 +13,7 @@ export type UserWithSettings = User & {
   subscription?: {
     expiresAt?: Date;
     offer: SubscriptionOffer;
+    subscribedAt?: Date;
     type: SubscriptionType;
   };
 };